+++ tests/test_envvar.py
@@ -81,7 +81,7 @@ def test_clean_envvar_id(env_name, clean_name):
         # (and show the defaults).
         (
             command,
-            "  --flag / --no-flag    [env var: "
+            "  --flag / --no-flag      [env var: "
             + ("CUSTOM, YO_FLAG" if os.name == "nt" else "custom, yo_FLAG")
             + "; default: no-flag]\n",
         ),